Intento de escritura numero 2
Cambio en el fichero EscribirXml.java y modificación de nombres
